[Store] Fast path for single-replica FreeRatioFirst allocation

+++ mooncake-store/src/allocator.cpp
@@ -126,7 +126,8 @@ std::unique_ptr<AllocatedBuffer> CachelibBufferAllocator::allocate(
     void* buffer = nullptr;
     try {
         // Allocate memory using CacheLib.
-        size_t padding_size = std::max(size, kMinSliceSize);
+        size_t padding_size =
+            std::max(size, static_cast<size_t>(kMinSliceSize));
         buffer = memory_allocator_->allocate(pool_id_, padding_size);
         if (!buffer) {
             VLOG(1) << "allocation_failed size=" << size
@@ -397,7 +398,8 @@ void* SimpleAllocator::allocate(size_t size) {
     }
 
     try {
-        size_t padding_size = std::max(size, kMinSliceSize);
+        size_t padding_size =
+            std::max(size, static_cast<size_t>(kMinSliceSize));
         void* ptr = memory_allocator_->allocate(pool_id_, padding_size);
         if (!ptr) {
             // This allocator is used in client side. Though the failure will
